pyproject.toml
[tool.poetry]
name = "langchain-learn"
version = "0.1.0"
description = ""
authors = ["bharathkreddy <reddybharath.k@gmail.com>"]
readme = "README.md"
packages = [{include = "langchain_learn"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.334"
openai = "^1.2.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
